refactor(js): removed tailwind content for js files and removed JSON5 and DOMPurify and docs with gatbsy 5
Now Xt.node doesn't sanitize html, set up your own Xt.sanitizeFnc for sanitize Xt.node
