+++ .github/workflows/deploy.yml
@@ -10,7 +10,7 @@ on:
       - completed
 
 env:
-  NODE: 16.x
+  NODE: 18.x
 
 jobs:
   deploy:
