+++ .github/workflows/test.yml
@@ -7,7 +7,7 @@ on:
     branches: ['*']
 
 env:
-  NODE: 16.x
+  NODE: 18.x
 
 jobs:
   codeql:
