+++ SETUP.md
@@ -51,8 +51,6 @@ module.exports = {
   presets: [
     require('tailwindcss/defaultConfig'), require('xtendui/tailwind.preset'),
   ],
-  // put other content e.g.: './src/**/*.{html,js}'
-  content: ['./node_modules/xtendui/src/*.mjs'],
 }
 ```
 
@@ -80,9 +78,7 @@ You can import the **components you need** as described in the demos.
 import { Xt } from 'xtendui'
 import 'xtendui/src/toggle'
 ```
-
-> Some bundlers doesn't support `package.json#exports` if you get module not found use the extensions `.mjs` (e.g. `import 'xtendui/src/toggle.mjs'`).
-
+ì
 #### Gsap
 
 Xtend UI uses [gsap](https://github.com/greensock/GSAP) for javascript animations in some demos, but it's not included in the library.
@@ -115,7 +111,7 @@ module.exports = {
 }
 ```
 
-Be sure to transpile `mjs` extension and exclude `node_modules` but include `xtendui`, for example with `babel-loader`:
+Be sure to transpile `js` extension and exclude `node_modules` but include `xtendui`, for example with `babel-loader`:
 
 ```js
 {
