+++ gatsby-config.js
@@ -9,7 +9,6 @@ require('dotenv').config()
 module.exports = {
   flags: {
     DEV_SSR: true,
-    //FAST_REFRESH: true,
   },
   siteMetadata: {
     siteUrl: 'https://xtendui.com',
