+++ package.json
@@ -17,27 +17,7 @@
     "vanilla-javascript",
     "tailwindcss"
   ],
-  "main": "src/xt.mjs",
-  "exports": {
-    ".": "./src/xt.mjs",
-    "./src/xt": "./src/xt.mjs",
-    "./tailwind.preset": "./tailwind.preset.js",
-    "./src/drop": "./src/drop.mjs",
-    "./src/form": "./src/form.mjs",
-    "./src/googlelocator": "./src/googlelocator.mjs",
-    "./src/groupnumber": "./src/groupnumber.mjs",
-    "./src/infinitescroll": "./src/infinitescroll.mjs",
-    "./src/mousefollow": "./src/mousefollow.mjs",
-    "./src/overlay": "./src/overlay.mjs",
-    "./src/ripple": "./src/ripple.mjs",
-    "./src/scrollto": "./src/scrollto.mjs",
-    "./src/slider": "./src/slider.mjs",
-    "./src/stickyflow": "./src/stickyflow.mjs",
-    "./src/textareaautosize": "./src/textareaautosize.mjs",
-    "./src/toggle": "./src/toggle.mjs",
-    "./src/tooltip": "./src/tooltip.mjs",
-    "./src/usability": "./src/usability.mjs"
-  },
+  "main": "src/xt.js",
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
   },
@@ -56,7 +36,7 @@
     "deploy": "gh-pages -d public",
     "dist": "yarn setup && cross-env DIST=true webpack --progress",
     "release": "npx semantic-release --no-ci --dry-run --debug",
-    "lint": "eslint . --ext .js,.mjs",
+    "lint": "eslint . --ext .js",
     "lint:generate": "eslint . --ext .jsx",
     "test": "start-server-and-test serve http://localhost:8080 test:run",
     "test:run": "cypress run",
@@ -65,84 +45,83 @@
     "postversion": "yarn dist"
   },
   "dependencies": {
-    "@popperjs/core": "^2.11.5",
-    "dompurify": "^2.3.6",
-    "focus-trap": "^6.9.0",
-    "json5": "^2.2.1",
+    "@popperjs/core": "^2.11.6",
+    "deepmerge": "^4.2.2",
+    "focus-trap": "^7.1.0",
     "lodash": "^4.17.21"
   },
   "peerDependencies": {
     "tailwindcss": "^3.0.12"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.3",
-    "@babel/preset-env": "^7.19.4",
+    "@babel/core": "^7.20.2",
+    "@babel/preset-env": "^7.20.2",
     "@hot-loader/react-dom": "^17.0.2",
     "@semantic-release/changelog": "^6.0.1",
     "@semantic-release/commit-analyzer": "^9.0.2",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.0.4",
+    "@semantic-release/github": "^8.0.6",
     "@semantic-release/npm": "^9.0.1",
     "@semantic-release/release-notes-generator": "^10.0.3",
-    "autoprefixer": "^10.4.7",
-    "babel-loader": "^8.2.5",
-    "babel-preset-gatsby": "^2.13.0",
+    "autoprefixer": "^10.4.13",
+    "babel-loader": "^9.1.0",
+    "babel-preset-gatsby": "^3.0.0",
     "clipboard": "^2.0.11",
-    "core-js": "^3.25.5",
+    "core-js": "^3.26.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
-    "cssnano": "^5.1.7",
-    "cypress": "^9.6.0",
-    "cypress-axe": "^0.14.0",
+    "css-loader": "^6.7.2",
+    "cssnano": "^5.1.14",
+    "cypress": "^11.0.1",
+    "cypress-axe": "^1.0.0",
     "cypress-plugin-tab": "^1.0.5",
-    "dompurify": "^2.3.1",
-    "dotenv": "^16.0.0",
-    "eslint": "8.14.0",
+    "dotenv": "^16.0.3",
+    "eslint": "8.27.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.10",
     "esm": "^3.2.25",
-    "gatsby": "^4.13.1",
-    "gatsby-plugin-catch-links": "^4.13.0",
-    "gatsby-plugin-image": "^2.13.0",
-    "gatsby-plugin-manifest": "^4.13.0",
+    "gatsby": "^5.0.1",
+    "gatsby-plugin-catch-links": "^5.0.0",
+    "gatsby-plugin-image": "^3.0.0",
+    "gatsby-plugin-manifest": "^5.0.0",
     "gatsby-plugin-module-resolver": "^1.0.3",
-    "gatsby-plugin-page-creator": "^4.13.0",
-    "gatsby-plugin-postcss": "^5.13.0",
-    "gatsby-plugin-react-helmet": "^5.13.0",
-    "gatsby-plugin-sitemap": "^5.13.0",
+    "gatsby-plugin-page-creator": "^5.0.0",
+    "gatsby-plugin-postcss": "^6.0.0",
+    "gatsby-plugin-react-helmet": "^6.0.0",
+    "gatsby-plugin-sitemap": "^6.0.0",
     "gatsby-remark-classes": "^1.0.2",
     "gatsby-remark-component": "^1.1.3",
     "gatsby-remark-external-links": "^0.0.4",
-    "gatsby-source-contentful": "^7.11.0",
-    "gatsby-source-filesystem": "^4.13.0",
-    "gatsby-transformer-remark": "^5.13.0",
-    "gh-pages": "^3.2.3",
+    "gatsby-source-contentful": "^8.0.0",
+    "gatsby-source-filesystem": "^5.0.0",
+    "gatsby-transformer-remark": "^6.0.0",
+    "gh-pages": "^4.0.0",
     "glob": "^8.0.1",
-    "gsap": "^3.11.0",
-    "mini-css-extract-plugin": "^2.6.0",
+    "gsap": "^3.11.3",
+    "mini-css-extract-plugin": "^2.6.1",
     "path": "^0.12.7",
-    "postcss": "^8.4.13",
-    "postcss-import": "^14.1.0",
-    "postcss-loader": "^6.2.1",
-    "prettier-eslint": "^14.0.2",
-    "prismjs": "^1.28.0",
+    "postcss": "^8.4.19",
+    "postcss-import": "^15.0.0",
+    "postcss-loader": "^7.0.1",
+    "prettier-eslint": "^15.0.1",
+    "prismjs": "^1.29.0",
     "prop-types": "^15.8.1",
-    "purgecss": "^4.1.3",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "purgecss": "^5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "regenerator-runtime": "^0.13.9",
+    "regenerator-runtime": "^0.13.10",
     "rehype-react": "^7.1.1",
     "remark-grid-tables": "^2.1.2",
-    "semantic-release": "^19.0.2",
+    "semantic-release": "^19.0.5",
     "start-server-and-test": "^1.14.0",
-    "tailwindcss": "^3.0.24",
-    "terser-webpack-plugin": "^5.3.1",
-    "uuid": "^8.3.2",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
-    "write": "^2.0.0"
+    "tailwindcss": "^3.2.4",
+    "terser-webpack-plugin": "^5.3.6",
+    "uuid": "^9.0.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0",
+    "write": "^2.0.0",
+    "xss": "^1.0.14"
   }
 }
