+++ src/gatsby/templates/includes/header.js
@@ -168,11 +168,7 @@ export default function Header({ page }) {
                         {page.menus.posts.map(({ post }, i) => (
                           <div
                             className={`gatsby_tooltip_outside_link ${
-                              (typeof window === 'undefined' || window.access !== 'admin') &&
-                              post.frontmatter.tags &&
-                              post.frontmatter.tags.includes('hidden')
-                                ? 'hidden'
-                                : ''
+                              post.frontmatter.tags && post.frontmatter.tags.includes('hidden') ? 'hidden' : ''
                             }`}
                             key={i}>
                             <Link
@@ -207,9 +203,7 @@ export default function Header({ page }) {
                                       post.frontmatter.parent === post.frontmatter.title ? (
                                         <div
                                           className={`gatsby_site-header_item_container ${
-                                            (typeof window === 'undefined' || window.access !== 'admin') &&
-                                            post.frontmatter.tags &&
-                                            post.frontmatter.tags.includes('hidden')
+                                            post.frontmatter.tags && post.frontmatter.tags.includes('hidden')
                                               ? 'hidden'
                                               : ''
                                           }`}
