+++ src/infinitescroll.js
@@ -4,9 +4,7 @@
  * @license MIT (https://github.com/xtendui/xtendui/blob/master/LICENSE.txt)
  */
 
-import { Xt } from './xt.mjs'
-import JSON5 from 'json5'
-Xt.JSON5 = JSON5
+import { Xt } from './xt.js'
 
 /**
  * Infinitescroll
@@ -664,7 +662,7 @@ if (typeof window !== 'undefined') {
       // vars
 
       const optionsMarkup = ref.getAttribute(`data-${Xt.Infinitescroll.componentName}`)
-      const options = optionsMarkup ? JSON5.parse(optionsMarkup) : {}
+      const options = optionsMarkup ? JSON.stringify(optionsMarkup) : {}
 
       // init
