+++ src/overlay.js
@@ -4,10 +4,8 @@
  * @license MIT (https://github.com/xtendui/xtendui/blob/master/LICENSE.txt)
  */
 
-import { Xt } from './xt.mjs'
-import './toggle.mjs'
-import JSON5 from 'json5'
-Xt.JSON5 = JSON5
+import { Xt } from './xt.js'
+import './toggle.js'
 
 /**
  * Overlay
@@ -96,7 +94,7 @@ if (typeof window !== 'undefined') {
       // vars
 
       const optionsMarkup = ref.getAttribute(`data-${Xt.Overlay.componentName}`)
-      const options = optionsMarkup ? JSON5.parse(optionsMarkup) : {}
+      const options = optionsMarkup ? JSON.stringify(optionsMarkup) : {}
 
       // init
