+++ src/ripple.js
@@ -4,14 +4,12 @@
  * @license MIT (https://github.com/xtendui/xtendui/blob/master/LICENSE.txt)
  */
 
-import { Xt } from './xt.mjs'
-import JSON5 from 'json5'
-Xt.JSON5 = JSON5
+import { Xt } from './xt.js'
 
 /**
- * Stickyflow
+ * Ripple
  */
-class Stickyflow {
+class Ripple {
   /**
    * fields
    */
@@ -19,15 +17,19 @@ class Stickyflow {
   _optionsDefault
   _optionsInitial
   _componentNs
-  _scrollTopOld
   componentName
   ns
   options
   initial
   disabled = false
   container
-  element
-  filler
+  inner
+  targets
+  size
+  top
+  left
+  sizeFinal
+  scaleFinal
 
   /**
    * constructor
@@ -76,17 +78,15 @@ class Stickyflow {
     Xt._initMatches({ self, optionsInitial: self._optionsInitial })
     // vars
     self.initial = true
-    // elements
-    self.element = self.container.querySelector(options.element)
-    self.filler = self.container.querySelector(options.filler)
-    // vars
-    self._scrollTopOld = 0
-    // events
-    const changeHandler = Xt.dataStorage.put(window, `scroll resize/${self.ns}`, self._eventChange.bind(self))
-    addEventListener('scroll', changeHandler)
-    addEventListener('resize', changeHandler)
-    // initial
-    self._initStart()
+    // inner
+    if (!self.inner) {
+      self.container.append(Xt.node({ str: '<div class="xt-ripple-inner"></div>' }))
+      self.inner = self.container.querySelector(':scope > .xt-ripple-inner')
+    }
+    // on
+    const onHandler = Xt.dataStorage.put(self.container, `mousedown touchstart/${self.ns}`, self._eventStart.bind(self))
+    self.container.addEventListener('mousedown', onHandler)
+    self.container.addEventListener('touchstart', onHandler, { passive: true })
     // init
     // needs frameDouble after ondone
     Xt.frameDouble({
@@ -111,75 +111,96 @@ class Stickyflow {
     }
   }
 
-  /**
-   * init start
-   */
-  _initStart() {
-    const self = this
-    // disabled
-    if (self.disabled) {
-      return
-    }
-    // initial
-    self._eventChange()
-  }
-
   //
   // methods
   //
 
   /**
-   * eventChange
-   * @param {Node|HTMLElement|EventTarget|Window} step
+   * eventStart
    * @param {Event} e
    */
-  _eventChange() {
+  _eventStart(e) {
+    const self = this
+    const options = self.options
+    // disabled
+    if (self.disabled) {
+      return
+    }
+    // check if inside onlyInside
+    if (!options.onlyInside || e.target.closest(options.onlyInside)) {
+      self.inner.append(Xt.node({ str: '<div class="xt-ripple"></div>' }))
+      // fix prevent dragging links and images
+      if (e.type === 'mousedown') {
+        e.preventDefault()
+      }
+      // size
+      const h = self.container.offsetHeight
+      const w = self.container.offsetWidth
+      const sizeObject = h > w ? h : w
+      const size = sizeObject * options.sizeInitial + 1 // at least 1 pixel
+      // scale from diagonal
+      const sizeFinal = Math.sqrt(Math.pow(h, 2) + Math.pow(w, 2))
+      const scaleFinal = (sizeFinal / size) * 2
+      // offset (using clientX and clientY for touch so we need to get values with getBoundingClientRect also)
+      let y
+      let x
+      if (e.clientX !== undefined) {
+        y = e.clientY
+        x = e.clientX
+      } else if (e.touches && e.touches.length) {
+        y = e.touches[0].clientY
+        x = e.touches[0].clientX
+      }
+      const rectTarget = e.target.getBoundingClientRect()
+      y = y - rectTarget.top
+      x = x - rectTarget.left
+      // position
+      if (self.container !== e.target) {
+        const rectObject = self.container.getBoundingClientRect()
+        y += rectTarget.top - rectObject.top
+        x += rectTarget.left - rectObject.left
+      }
+      const top = y - size / 2
+      const left = x - size / 2
+      // dispatch event
+      self.size = size
+      self.top = top
+      self.left = left
+      self.sizeFinal = sizeFinal
+      self.scaleFinal = scaleFinal
+      // dispatch event
+      self.container.dispatchEvent(
+        new CustomEvent(`on.${self._componentNs}`, {
+          detail: e,
+        })
+      )
+      // off
+      const endHandler = Xt.dataStorage.put(window, `mouseup touchend/${self.ns}`, self._eventEnd.bind(self))
+      addEventListener('mouseup', endHandler)
+      addEventListener('touchend', endHandler, { passive: true })
+    }
+  }
+
+  /**
+   * eventEnd
+   * @param {Event} e
+   */
+  _eventEnd(e) {
     const self = this
     // disabled
     if (self.disabled) {
       return
     }
-    // position
-    const scrollTop = document.scrollingElement.scrollTop
-    const windowHeight = window.innerHeight
-    const objectHeight = self.element.offsetHeight
-    if (objectHeight < windowHeight) {
-      self.filler.style.height = ''
-      self.element.style.top = '0'
-      self.element.style.bottom = ''
-    } else {
-      if (scrollTop > self._scrollTopOld) {
-        if (!self.element.classList.contains('xt-stickyflow-top')) {
-          const pos = windowHeight - objectHeight
-          const height = Math.max(0, self.element.offsetTop - self.filler.offsetTop)
-          self.filler.style.height = `${height}px`
-          self.element.style.top = `${pos}px`
-          self.element.style.bottom = ''
-          self.element.classList.remove('xt-stickyflow-bottom')
-          self.element.classList.add('xt-stickyflow-top')
-        }
-      } else {
-        if (!self.element.classList.contains('xt-stickyflow-bottom')) {
-          const pos = windowHeight - objectHeight
-          const height = Math.max(0, self.element.offsetTop - self.filler.offsetTop)
-          self.filler.style.height = `${height}px`
-          self.element.style.top = ''
-          self.element.style.bottom = `${pos}px`
-          self.element.classList.add('xt-stickyflow-bottom')
-          self.element.classList.remove('xt-stickyflow-top')
-        }
-      }
-    }
-    // change
-    Xt.frame({
-      el: self.container,
-      ns: `${self.ns}Change`,
-      func: () => {
-        // dispatch event
-        self.container.dispatchEvent(new CustomEvent(`change.${self._componentNs}`))
-      },
-    })
-    self._scrollTopOld = scrollTop
+    // off
+    const endHandler = Xt.dataStorage.get(window, `mouseup touchend/${self.ns}`)
+    removeEventListener('mouseup', endHandler)
+    removeEventListener('touchend', endHandler)
+    // dispatch event
+    self.container.dispatchEvent(
+      new CustomEvent(`off.${self._componentNs}`, {
+        detail: e,
+      })
+    )
   }
 
   //
@@ -209,12 +230,6 @@ class Stickyflow {
     if (!self.disabled) {
       // disable
       self.disabled = true
-      // position
-      self.filler.style.height = ''
-      self.element.style.top = ''
-      self.element.style.bottom = ''
-      self.element.classList.remove('xt-stickyflow-bottom')
-      self.element.classList.remove('xt-stickyflow-top')
       // dispatch event
       if (!skipEvent) {
         self.container.dispatchEvent(new CustomEvent(`status.${self._componentNs}`))
@@ -240,10 +255,18 @@ class Stickyflow {
    */
   destroy() {
     const self = this
+    // inner
+    self.inner.remove()
+    self.inner = null
     // remove events
-    const changeHandler = Xt.dataStorage.get(window, `scroll resize/${self.ns}`)
-    removeEventListener('scroll', changeHandler)
-    removeEventListener('resize', changeHandler)
+    // on
+    const onHandler = Xt.dataStorage.get(self.container, `mousedown touchstart/${self.ns}`)
+    self.container.removeEventListener('mousedown', onHandler)
+    self.container.removeEventListener('touchstart', onHandler, { passive: true })
+    // off
+    const endHandler = Xt.dataStorage.get(window, `mouseup touchend/${self.ns}`)
+    removeEventListener('mouseup', endHandler)
+    removeEventListener('touchend', endHandler)
     // initialized class
     self.container.removeAttribute(`data-${self.componentName}-init`)
     // set self
@@ -261,43 +284,16 @@ class Stickyflow {
 // options
 //
 
-Stickyflow.componentName = 'xt-stickyflow'
-Stickyflow.optionsDefault = {
+Ripple.componentName = 'xt-ripple'
+Ripple.optionsDefault = {
   debug: false,
-  // elements
-  element: false,
-  filler: false,
+  // ripple
+  sizeInitial: 0.3,
+  onlyInside: 'a, button, .xt-button',
 }
 
 //
 // export
 //
 
-Xt.Stickyflow = Stickyflow
-
-//
-// observe
-//
-
-if (typeof window !== 'undefined') {
-  Xt.mount({
-    matches: `[data-${Xt.Stickyflow.componentName}]`,
-    mount: ({ ref }) => {
-      // vars
-
-      const optionsMarkup = ref.getAttribute(`data-${Xt.Stickyflow.componentName}`)
-      const options = optionsMarkup ? JSON5.parse(optionsMarkup) : {}
-
-      // init
-
-      let self = new Xt.Stickyflow(ref, options)
-
-      // unmount
-
-      return () => {
-        self.destroy()
-        self = null
-      }
-    },
-  })
-}
+Xt.Ripple = Ripple
