+++ src/stickyflow.js
@@ -4,14 +4,12 @@
  * @license MIT (https://github.com/xtendui/xtendui/blob/master/LICENSE.txt)
  */
 
-import { Xt } from './xt.mjs'
-import JSON5 from 'json5'
-Xt.JSON5 = JSON5
+import { Xt } from './xt.js'
 
 /**
- * Textareaautosize
+ * Stickyflow
  */
-class Textareaautosize {
+class Stickyflow {
   /**
    * fields
    */
@@ -19,13 +17,15 @@ class Textareaautosize {
   _optionsDefault
   _optionsInitial
   _componentNs
+  _scrollTopOld
   componentName
   ns
   options
   initial
   disabled = false
   container
-  form
+  element
+  filler
 
   /**
    * constructor
@@ -74,18 +74,15 @@ class Textareaautosize {
     Xt._initMatches({ self, optionsInitial: self._optionsInitial })
     // vars
     self.initial = true
-    // key
-    const changeHandler = Xt.dataStorage.put(
-      self.container,
-      `keydown keyup reset/${self.ns}`,
-      self._keychange.bind(self)
-    )
-    self.container.addEventListener('keydown', changeHandler)
-    self.container.addEventListener('keyup', changeHandler)
-    self.form = self.container.closest('form')
-    if (self.form) {
-      self.form.addEventListener('reset', changeHandler)
-    }
+    // elements
+    self.element = self.container.querySelector(options.element)
+    self.filler = self.container.querySelector(options.filler)
+    // vars
+    self._scrollTopOld = 0
+    // events
+    const changeHandler = Xt.dataStorage.put(window, `scroll resize/${self.ns}`, self._eventChange.bind(self))
+    addEventListener('scroll', changeHandler)
+    addEventListener('resize', changeHandler)
     // initial
     self._initStart()
     // init
@@ -121,8 +118,8 @@ class Textareaautosize {
     if (self.disabled) {
       return
     }
-    // logic
-    self._keychange.bind(self)()
+    // initial
+    self._eventChange()
   }
 
   //
@@ -130,25 +127,57 @@ class Textareaautosize {
   //
 
   /**
-   * keychange
+   * eventChange
+   * @param {Node|HTMLElement|EventTarget|Window} step
+   * @param {Event} e
    */
-  _keychange() {
+  _eventChange() {
     const self = this
     // disabled
     if (self.disabled) {
       return
     }
-    // size
-    self.container.style.height = '5px'
-    self.container.style.height = `${self.container.scrollHeight}px` // fixes both safari RAF and form reset
+    // position
+    const scrollTop = document.scrollingElement.scrollTop
+    const windowHeight = window.innerHeight
+    const objectHeight = self.element.offsetHeight
+    if (objectHeight < windowHeight) {
+      self.filler.style.height = ''
+      self.element.style.top = '0'
+      self.element.style.bottom = ''
+    } else {
+      if (scrollTop > self._scrollTopOld) {
+        if (!self.element.classList.contains('xt-stickyflow-top')) {
+          const pos = windowHeight - objectHeight
+          const height = Math.max(0, self.element.offsetTop - self.filler.offsetTop)
+          self.filler.style.height = `${height}px`
+          self.element.style.top = `${pos}px`
+          self.element.style.bottom = ''
+          self.element.classList.remove('xt-stickyflow-bottom')
+          self.element.classList.add('xt-stickyflow-top')
+        }
+      } else {
+        if (!self.element.classList.contains('xt-stickyflow-bottom')) {
+          const pos = windowHeight - objectHeight
+          const height = Math.max(0, self.element.offsetTop - self.filler.offsetTop)
+          self.filler.style.height = `${height}px`
+          self.element.style.top = ''
+          self.element.style.bottom = `${pos}px`
+          self.element.classList.add('xt-stickyflow-bottom')
+          self.element.classList.remove('xt-stickyflow-top')
+        }
+      }
+    }
+    // change
     Xt.frame({
       el: self.container,
-      ns: `${self.ns}Keychange`,
+      ns: `${self.ns}Change`,
       func: () => {
-        self.container.style.height = '5px' // fixes both safari RAF and form reset
-        self.container.style.height = `${self.container.scrollHeight}px`
+        // dispatch event
+        self.container.dispatchEvent(new CustomEvent(`change.${self._componentNs}`))
       },
     })
+    self._scrollTopOld = scrollTop
   }
 
   //
@@ -178,12 +207,12 @@ class Textareaautosize {
     if (!self.disabled) {
       // disable
       self.disabled = true
-      // size
-      self.container.style.height = ''
-      Xt.frame({
-        el: self.container,
-        ns: `${self.ns}Keychange`,
-      })
+      // position
+      self.filler.style.height = ''
+      self.element.style.top = ''
+      self.element.style.bottom = ''
+      self.element.classList.remove('xt-stickyflow-bottom')
+      self.element.classList.remove('xt-stickyflow-top')
       // dispatch event
       if (!skipEvent) {
         self.container.dispatchEvent(new CustomEvent(`status.${self._componentNs}`))
@@ -209,15 +238,10 @@ class Textareaautosize {
    */
   destroy() {
     const self = this
-    // reset
-    self.container.style.height = ''
     // remove events
-    const changeHandler = Xt.dataStorage.get(self.container, `keydown keyup reset/${self.ns}`)
-    self.container.removeEventListener('keydown', changeHandler)
-    self.container.removeEventListener('keyup', changeHandler)
-    if (self.form) {
-      self.form.removeEventListener('reset', changeHandler)
-    }
+    const changeHandler = Xt.dataStorage.get(window, `scroll resize/${self.ns}`)
+    removeEventListener('scroll', changeHandler)
+    removeEventListener('resize', changeHandler)
     // initialized class
     self.container.removeAttribute(`data-${self.componentName}-init`)
     // set self
@@ -235,16 +259,19 @@ class Textareaautosize {
 // options
 //
 
-Textareaautosize.componentName = 'xt-textareaautosize'
-Textareaautosize.optionsDefault = {
+Stickyflow.componentName = 'xt-stickyflow'
+Stickyflow.optionsDefault = {
   debug: false,
+  // elements
+  element: false,
+  filler: false,
 }
 
 //
 // export
 //
 
-Xt.Textareaautosize = Textareaautosize
+Xt.Stickyflow = Stickyflow
 
 //
 // observe
@@ -252,16 +279,16 @@ Xt.Textareaautosize = Textareaautosize
 
 if (typeof window !== 'undefined') {
   Xt.mount({
-    matches: `[data-${Xt.Textareaautosize.componentName}]`,
+    matches: `[data-${Xt.Stickyflow.componentName}]`,
     mount: ({ ref }) => {
       // vars
 
-      const optionsMarkup = ref.getAttribute(`data-${Xt.Textareaautosize.componentName}`)
-      const options = optionsMarkup ? JSON5.parse(optionsMarkup) : {}
+      const optionsMarkup = ref.getAttribute(`data-${Xt.Stickyflow.componentName}`)
+      const options = optionsMarkup ? JSON.stringify(optionsMarkup) : {}
 
       // init
 
-      let self = new Xt.Textareaautosize(ref, options)
+      let self = new Xt.Stickyflow(ref, options)
 
       // unmount
